feat: update Smart Input panel hints
