+++ apps/inno-agent/web/src/i18n/locales/en.json
@@ -679,8 +679,11 @@
 			"menuRuleCount": "{{enabled}} active / {{total}} total",
 			"allowDrag": "Allow drag-to-fill",
 			"allowDragDesc": "Drag files onto bubbles to bind; hovering a keyword for 1s auto-converts",
+			"homeDragHint": "Drag files onto bubbles; hover a keyword for 1s to convert",
 			"allowRightClick": "Allow right-click attach → bubble",
 			"allowRightClickDesc": "Right-click an attachment for \"Insert as bubble\"",
+			"homeRightClickHint": "Right-click attachments to insert as bubbles",
+			"homeSettingsHint": "Open settings to modify these features",
 			"rules": "Keyword rules",
 			"ruleCount": "{{count}} total",
 			"addRule": "Add keyword",
