+++ apps/inno-agent/web/src/i18n/locales/zh-CN.json
@@ -679,8 +679,11 @@
 			"menuRuleCount": "已开启 {{enabled}} 条 / 共 {{total}} 条",
 			"allowDrag": "允许拖入填充",
 			"allowDragDesc": "拖文件到气泡绑定；拖文件悬停关键词 1 秒自动转换",
+			"homeDragHint": "拖文件到气泡绑定，悬停关键词 1 秒自动转换",
 			"allowRightClick": "允许右键附件转气泡",
 			"allowRightClickDesc": "附件右键「插入为气泡」",
+			"homeRightClickHint": "右键附件插入为气泡",
+			"homeSettingsHint": "可进入设置修改上述功能",
 			"rules": "关键词规则",
 			"ruleCount": "共 {{count}} 条",
 			"addRule": "新增关键词",
